+++ artifacts/burger-gn/src/App.tsx
@@ -18,6 +18,7 @@ import ClubeCliente from "./pages/ClubeCliente";
 import { MyOrderFab } from "./components/MyOrderFab";
 import { AdminPanelScroll } from "./components/AdminPanelScroll";
 import { AdminNotificationEngine } from "./components/AdminNotificationEngine";
+import { PrintAgentGuard } from "./components/PrintAgentGuard";
 import AdminLogin from "./pages/admin/Login";
 import AdminForgotPassword from "./pages/admin/ForgotPassword";
 import AdminSecurity from "./pages/admin/Security";
@@ -73,6 +74,7 @@ function ProtectedAdminRoute({ component: Component }: { component: React.Compon
     <>
       <AdminPanelScroll />
       <AdminNotificationEngine />
+      <PrintAgentGuard />
       <Component />
     </>
   );